requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.7.1
certifi==2019.6.16
chardet==3.0.4
idna==2.8
lxml==4.3.4
pymongo==3.8.0
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
six==1.13.0
soupsieve==1.9.1
tzlocal==2.0.0
urllib3==1.25.3
